requirements.txt
# API
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

# HTTP (async Ollama çağrıları)
httpx==0.27.2

# PDF işleme
pdfplumber==0.11.4

# Chunking
langchain-text-splitters==0.3.0

# Vektör DB
qdrant-client==1.11.3

# LlamaIndex orkestrasyon
llama-index-core==0.11.20
llama-index-llms-ollama==0.3.6
llama-index-embeddings-ollama==0.3.1

# Evaluation
ragas==0.2.6
datasets==3.0.1

# Yardımcı
python-dotenv==1.0.1
